requirements.txt
# Core dependencies
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
python-dotenv>=1.0.0

# Music format handling
music21>=9.1.0
mido>=1.3.0

# File handling and utilities
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# CLI and user interaction
click>=8.1.0
colorama>=0.4.6
